requirements.txt
cloudpickle
grpcio>=1.11.0
numpy>=1.14.1
Pillow>=4.2.1
protobuf==3.19.0
pyyaml>=3.1.0
pybullet
gym==0.21.0
#stable-baselines3
opencv-python